Update docs (#237)
* Update index.md

Updated links from public MM to Matrix

* Update outdated commands in tutorial

Updated the `juju run-action` to `juju run`. I removed the `--wait` flag as it no longer works without an argument; I am not opposed to adding `--wait` back in, but I would need guidance on an appropriate value. 

Updated `juju destroy model` to `juju destroy-model`, changed `--release-storage` to `--destroy-storage` and removed the `-y` flag.

* Update integrate-with-cos.md

Updated `juju run-action` to `juju run` and removed the `--wait` flag as it requires an argument. I'm not opposed to adding `--wait` back in but would need guidance on how long the reader should wait for.

* Capitalise Grafana in integrate-with-cos.md

* Update enable-waf.md

Formatted the notice about the Modsecurity feature availability as a note.

* Update configure-hostname.md

Remove a colon

* Update configure-initial-settings.md

Reformatted the notice about configuring initial settings as a note.

* Update retrieve-initial-credentials.md

Changed `juju run-action` to `juju run` and removed the `--wait` flag. I'm not opposed to adding `--wait` back in but would need guidance on what argument to use. 

Also reformatted the note at the end as a note and modified the grammar slightly.

* Update configure-object-storage.md

Removed an incomplete sentence. If the sentence was necessary, let me know how to complete it.

* Update rotate-secrets.md

Changed `juju run-action` to `juju run` and removed the `-wait` flag. I'm not opposed to adding `--wait` back in but would need guidance on what the argument should be.

* Update plugins.md

it's --> its

* Update lifecycle-events.md

Capitalise Juju and Pebble.

Update the upgrade-charm section to not consist of fragmented sentences.

* Update oci-image.md

Add a word

* Update relations.md

Format `loki_push_api` as code; add a word